LA-135: Map dtos for the ms-selection service

+++ src/main/java/com/selina/lending/internal/dto/OfferDto.java
@@ -31,7 +31,7 @@ public class OfferDto {
     Double aprc;
     Double cltv;
     Double offerBalance;
-    Double initialTerm;
+    Integer initialTerm;
     Double initialRate;
     Double brokerFeesIncluded;
     Double reversionTerm;
